@@ -3,24 +3,21 @@
 namespace Caloriscz\Menus\Admin;
 
 use Nette\Application\UI\Control;
-use Nette\Database\Context;
+use Nette\Database\Explorer;
 
 class MainMenuControl extends Control
 {
-    /** @var Context @inject */
-    public $database;
+    /** @var Explorer @inject */
+    public Explorer $database;
 
-    public function __construct(Context $database)
+    public function __construct(Explorer $database)
     {
         $this->database = $database;
     }
 
     public function render(): void
     {
-        if (isset($this->getPresenter()->template->member)) {
-            $this->template->member = $this->getPresenter()->template->member;
-        }
-
+        $this->template->member = $this->getPresenter()->template->member;
         $this->template->settings = $this->getPresenter()->template->settings;
         $this->template->pageTypes = $this->database->table('pages_types');
 
