@@ -4,15 +4,14 @@ namespace Caloriscz\Menus;
 
 use App\Forms\Menu\InsertMenuControl;
 use Nette\Application\UI\Control;
-use Nette\Database\Context;
+use Nette\Database\Explorer;
 use Tracy\Debugger;
 
 class MenuListControl extends Control
 {
-    /** @var Context */
-    public $database;
+    public Explorer $database;
 
-    public function __construct(Context $database)
+    public function __construct(Explorer $database)
     {
         $this->database = $database;
     }
@@ -29,7 +28,6 @@ class MenuListControl extends Control
     public function handleDelete(): void
     {
         $this->database->table('menu_menus')->get($this->getPresenter()->getParameter('id'))->delete();
-
         $this->getPresenter()->redirect('this');
     }
 
