@@ -4,16 +4,15 @@ namespace Caloriscz\Settings;
 
 use Caloriscz\Utilities\PagingControl;
 use Nette\Application\UI\Control;
-use Nette\Database\Context;
+use Nette\Database\Explorer;
 use Nette\Utils\Paginator;
 
 class BlackListControl extends Control
 {
 
-    /** @var Context */
-    public $database;
+    public Explorer $database;
 
-    public function __construct(Context $database)
+    public function __construct(Explorer $database)
     {
         $this->database = $database;
     }
@@ -22,7 +21,7 @@ class BlackListControl extends Control
      * @param $id
      * @throws \Nette\Application\AbortException
      */
-    public function handleDelete($id): void
+    public function handleDelete(int $id): void
     {
         $this->database->table('blacklist')->get($id)->delete();
         $this->presenter->redirect('this', ['id' => $this->presenter->getParameter('id')]);
@@ -35,10 +34,10 @@ class BlackListControl extends Control
 
     /**
      *
-     * @param type $id
-     * @param type $type Insert blaclist word or sentence
+     * @param int $id
+     * @param int $type Insert blaclist word or sentence
      */
-    public function render($id = null, $type = null): void
+    public function render(int $id = null, int $type = null): void
     {
         $this->template->type = $type;
         $blacklist = $this->database->table('blacklist')->order('title');
@@ -46,7 +45,7 @@ class BlackListControl extends Control
         $paginator = new Paginator();
         $paginator->setItemCount($blacklist->count('*'));
         $paginator->setItemsPerPage(20);
-        $paginator->setPage($this->presenter->getParameter('page'));
+        $paginator->setPage($this->presenter->getParameter('page') ?? 1);
 
         $this->template->blacklist = $blacklist->limit($paginator->getLength(), $paginator->getOffset());
         $this->template->paginator = $paginator;
