@@ -3,7 +3,7 @@
 namespace Caloriscz\Snippets;
 
 use Nette\Application\UI\Control;
-use Nette\Database\Context;
+use Nette\Database\Explorer;
 
 /**
  * Class SnippetControl
@@ -12,10 +12,9 @@ use Nette\Database\Context;
 class SnippetControl extends Control
 {
 
-    /** @var Context */
-    public $database;
+    public Explorer $database;
 
-    public function __construct(Context $database)
+    public function __construct(Explorer $database)
     {
         $this->database = $database;
     }
