@@ -3,15 +3,15 @@
 namespace Caloriscz\Menus;
 
 use Nette\Application\UI\Control;
-use Nette\Database\Context;
+use Nette\Database\Explorer;
 
 class MenuControl extends Control
 {
 
-    /** @var Context */
+    /** @var Explorer */
     public $database;
 
-    public function __construct(Context $database)
+    public function __construct(Explorer $database)
     {
         $this->database = $database;
     }
