@@ -5,6 +5,7 @@ namespace App\Forms\Pages;
 use App\Model\Document;
 use Nette\Application\UI\Control;
 use Nette\Database\Context;
+use Nette\Database\Explorer;
 use Nette\Forms\BootstrapUIForm;
 
 /**
@@ -19,9 +20,8 @@ class EditorControl extends Control
     /** @var Context */
     public $database;
 
-    public function __construct(Context $database)
+    public function __construct(Explorer $database)
     {
-        parent::__construct();
         $this->database = $database;
 
         $config = \HTMLPurifier_Config::createDefault();
